fix(core): shrink config and types debt below grandfathered ratchets (#2061)
