+++ packages/plugin-openclaw/openclaw.plugin.json
@@ -6281,6 +6281,57 @@
         "type": "string",
         "description": "Override workspace directory path"
       },
+      "externalWikis": {
+        "type": "array",
+        "default": [],
+        "description": "External compiled-wiki roots available only to dedicated on-demand search.",
+        "items": {
+          "type": "object",
+          "additionalProperties": false,
+          "properties": {
+            "id": {
+              "type": "string",
+              "pattern": "^[a-z0-9][a-z0-9_-]{0,63}$",
+              "description": "Stable wiki identifier."
+            },
+            "rootDir": {
+              "type": "string",
+              "description": "Absolute path or ~/ path to the external wiki root."
+            },
+            "enabled": {
+              "type": "boolean",
+              "default": true
+            },
+            "label": {
+              "type": "string"
+            },
+            "pagesDir": {
+              "type": "string",
+              "default": "wiki",
+              "description": "Wiki-root-relative concept-page directory."
+            },
+            "indexFile": {
+              "type": "string",
+              "default": "INDEX.md",
+              "description": "Wiki-root-relative catalog path."
+            },
+            "indexInQmd": {
+              "type": "boolean",
+              "default": false,
+              "description": "Index the corpus in its dedicated external-wiki collection."
+            },
+            "includeInDefaultRecall": {
+              "type": "boolean",
+              "default": false,
+              "description": "Reserved safety guard. Remnic rejects true."
+            }
+          },
+          "required": [
+            "id",
+            "rootDir"
+          ]
+        }
+      },
       "extractionBreakerAuthCooldownMs": {
         "type": "number",
         "default": 1800000,
