+++ packages/remnic-core/src/external-wiki-config.ts
@@ -1,6 +1,16 @@
 import path from "node:path";
-import type { ExternalWikiRoot } from "./types.js";
 import { expandTildePath } from "./utils/path.js";
+
+export interface ExternalWikiRoot {
+  id: string;
+  rootDir: string;
+  enabled: boolean;
+  label?: string;
+  pagesDir: string;
+  indexFile: string;
+  indexInQmd: boolean;
+  includeInDefaultRecall: false;
+}
 
 const EXTERNAL_WIKI_ID_PATTERN = /^[a-z0-9][a-z0-9_-]{0,63}$/;
