+++ packages/remnic-core/src/types.ts
@@ -4,6 +4,7 @@ import type { ActivityConfig } from "./activity/types.js";
 import type { WearablesConfig } from "./wearables/types.js";
 import type { ExtractionLivenessConfig } from "./extraction-liveness.js";
 import type { ReplicaPeersConfig } from "./replica-peers-config.js";
+import type { ExternalWikiRoot } from "./external-wiki-config.js";
 
 import type { BufferTurnOwner } from "./buffer-turn-helpers.js";
 export type ReasoningEffort = "none" | "low" | "medium" | "high";
@@ -670,17 +671,6 @@ export interface SemanticChunkingConfigShape {
   fallbackToRecursive: boolean;
 }
 
-export interface ExternalWikiRoot {
-  id: string;
-  rootDir: string;
-  enabled: boolean;
-  label?: string;
-  pagesDir: string;
-  indexFile: string;
-  indexInQmd: boolean;
-  includeInDefaultRecall: false;
-}
-
 export interface PluginConfig extends BoundedJsonlStateConfig {
   openaiApiKey: string | undefined;
   openaiBaseUrl: string | undefined;
@@ -744,7 +734,6 @@ export interface PluginConfig extends BoundedJsonlStateConfig {
   memoryOsPreset?: MemoryOsPresetName;
   qmdEnabled: boolean;
   qmdCollection: string;
-  /** External wiki collections are excluded from ordinary recall. V1 rejects true. */
   wikiMergeIntoRecall: boolean;
   qmdMaxResults: number;
   qmdEmbeddingBacklogThreshold: number;
@@ -764,11 +753,8 @@ export interface PluginConfig extends BoundedJsonlStateConfig {
   qmdChunkStrategy: "auto" | "regex";
   qmdCandidateLimit?: number;
   qmdQueryRerankEnabled: boolean;
-  // Daemon search plan; "hybrid" (default) = lex+vec+hyde. Issue #1335.
   qmdSearchStrategy: "hybrid" | "lex-vec" | "lex";
-  // Subprocess fallback command; "query" (default) keeps LLM expansion. Issue #1335.
   qmdSubprocessStrategy: "query" | "search";
-  // Per-call daemon search timeout in ms; default 8000. Issue #1335.
   qmdDaemonTimeoutMs: number;
   qmdIndexName?: string;
   qmdForceCpu: boolean;
