+++ scripts/config-contract/parsed-keys.snapshot.json
@@ -346,6 +346,7 @@
     "explicitCueRecallEnabled",
     "explicitCueRecallMaxChars",
     "explicitCueRecallMaxReferences",
+    "externalWikis",
     "extractionBreakerAuthCooldownMs",
     "extractionBreakerCooldownMs",
     "extractionBreakerFailureThreshold",
@@ -1182,6 +1183,7 @@
     "explicitCueRecallEnabled.trim",
     "explicitCueRecallMaxChars.trim",
     "explicitCueRecallMaxReferences.trim",
+    "externalWikis.map",
     "extractionFaithfulnessContextChars.trim",
     "extractionFaithfulnessGate.trim",
     "extractionFaithfulnessLocalParseFallback.trim",
